just removed validation

+++ app/Http/Controllers/ArticleController.php
@@ -23,43 +23,47 @@ class ArticleController extends Controller
     public function store(Request $request)
     {
 
-        $validatedData = $request->validate([
-            'title' => 'required|max:30',
-            'content' => 'required',
-            'author' => 'required',
-            'category' => 'required',
+        // $validatedData = $request->validate([
+        //     'title' => 'required|max:30',
+        //     'content' => 'required',
+        //     'author' => 'required',
+        //     'category' => 'required',
 
-        ]);
+        // ]);
         $article = new Article();
-        $article->title = $validatedData['title'];
-        $article->content = $validatedData['content'];
-        $article->author = $validatedData['author'];
-        $article->category = $validatedData['category'];
+        $article->title = $request['title'];
+        $article->content = $request['content'];
+        $article->author = $request['author'];
+        $article->category = $request['category'];
         $article->save();
         return response()->json($article, 201);
     }
     //
     public function update(Request $request, $id)
     {
-        $validatedData = $request->validate([
-            'title' => 'required|max:30',
-            'content' => 'required',
-            'author' => 'required',
-            'category' => 'required',
-        ]);
+        // $validatedData = $request->validate([
+        //     'title' => 'required|max:30',
+        //     'content' => 'required',
+        //     'author' => 'required',
+        //     'category' => 'required',
+        // ]);
         $article = Article::findOrFail($id);
-        $article->title = $validatedData['title'];
-        $article->content = $validatedData['content'];
-        $article->author = $validatedData['author'];
-        $article->category = $validatedData['category'];
+        $article->title = $request['title'];
+        $article->content = $request['content'];
+        $article->author = $request['author'];
+        $article->category = $request['category'];
         $article->save();
         return response()->json($article);
     }
     //
     public function destroy($id)
     {
-        $article = Article::findOrFail($id);
-        $article->delete();
-        return response()->json('Article deleted');
+        $article = Article::find($id);
+        if ($article) {
+            $article->delete();
+            return response()->json(['message' => 'Article deleted successfully']);
+        } else {
+            return response()->json(['message' => 'Article not found'], 404);
+        }
     }
 }
